+++ backend/routes/auth.js
@@ -1,5 +1,5 @@
 const express = require("express");
-const db = require("./db");
+const db = require("../db");
 const jwt = require('jsonwebtoken');
