+++ src/pages/components/statCards/mp_series.js
@@ -24,10 +24,10 @@ function MPSeries(props) {
   
     const fetchLibraries = () => {
       if (config) {
-        const url = `/stats/getMostPopularSeries`;
+        const url = `/stats/getMostPopularByType`;
        
         axios
-          .post(url, { days: props.days }, {
+          .post(url, { days: props.days, type:'Series' }, {
             headers: {
               Authorization: `Bearer ${config.token}`,
               "Content-Type": "application/json",
