+++ src/setupProxy.js
@@ -37,7 +37,7 @@ module.exports = function(app) {
     })
   );
   app.use(
-    `/data`,
+    `/backup`,
     createProxyMiddleware({
       target: `http://127.0.0.1:${process.env.PORT || 3003}`,
       changeOrigin: true,
